double

+++ src/WebApp/Model/OrderModel/OrderModel.cs
@@ -5,9 +5,9 @@ public class OrderModel
     public string UserName { get; set; }
     public string CoinName { get; set; }
     public int CoinId { get; set; }
-    public decimal PriceCoin { get; set; }
-    public decimal Amount { get; set; }
-    public decimal TotalPrice { get; set; }
+    public double PriceCoin { get; set; }
+    public double Amount { get; set; }
+    public double TotalPrice { get; set; }
     public DateTime? DateTime { get; set; }
 
     // BillingAddress
